maruchan/requirements.txt
discord.py==0.16.12
docker==2.2.1
